[nextest-runner] move status level code to another file
Not much code here at the moment, but a lot of tests that can live here.
