+++ nextest-runner/src/reporter/displayer/mod.rs
@@ -6,7 +6,9 @@
 mod formatters;
 mod imp;
 mod progress;
+mod status_level;
 mod unit_output;
 
 pub(crate) use imp::*;
+pub use status_level::*;
 pub use unit_output::*;
